Move background modes guard to location manager wrapper
The `allowsBackgroundLocationUpdates` API is not safe to call unless the host application has set location as a background mode. This guard was in place but was done at the level above where it should have been.

This change moves the guard to the MMECLLocationManagerWrapper that actually holds the real CLLocationManager instance. As the source of truth for all location manager instance interactions, it is the safest place to perform this check.