Introduce build system
This adds the start of a build system to this project. For now, it is contained in a new package.sh script that has several functions including the ability to create a namespaced header file that can be used in client projects to effecively make a uniquely named version of this library at compile time. The package script can be called from make because of a new Makefile that has also been added.

This also formally renames reachability in this project to account for its actual symbol name.