Remove dependency on SDK bundle
We used this only for localizable strings that were not localized. English is enough for error messages related to network failures of our own telemetry posts.

Removing the dependency on the bundle reduces the complexity of some distribution use cases like when we embed this as a library in our Unity SDK.