Make access token setable
Previously an access token could only be set once. This changes things so that a new access token value can be set on the events manager and that value is passed through to the api manager where it is actually used.

This makes it so that clients of the library can more easily and explicitly set alternative access tokens without having to hand off via user defaults.