Make base URL settable
This adds a new public API that makes it easy to set the base URL for where events will be posted.

It was previously possible to override the default URL with a value for MMETelemetryTestServerURL in user defaults. This replaces that backdoor option with a public property on events manager that passes through to the api client where the value is actually used.

Validation for a secure protocol remains but word “test” is not used since this API could be used to point to any alternative backend stack, not just the test server.

Host apps and SDKs can still use user defaults to discover, set, and store the URL for events posting and pass the value through with this new API.