Remove user defaults hooks for staging testing
This cleans up some references to configuration keys (MMETelemetryTestServerURL and MMETelemetryStagingAccessToken) that used as a backdoor to change the server and access token for testing. There are now actual APIs that replace the old user defaults approach.

This also removes the accessToken getter in API client since that object now always expects to be configured with an access token instead of sometimes pawing through user defaults to find one.