Add build system and namespace header generation (#27)
* Introduce build system

This adds the start of a build system to this project. For now, it is contained in a new package.sh script that has several functions including the ability to create a namespaced header file that can be used in client projects to effecively make a uniquely named version of this library at compile time. The package script can be called from make because of a new Makefile that has also been added.

This also formally renames reachability in this project to account for its actual symbol name.

* Update build system to replace checked in name header

This updates the name-header make command so that whenever a new name header is generated it updates the existing checked in one in the project.

* Fix bug where low power mode was always used for non-sim disabled case

Previously, the low power mode check always determined if the events manager was enabled or disabled. After this change, it is correctly used as only part of the overall check.

* Remove dependency on SDK bundle

We used this only for localizable strings that were not localized. English is enough for error messages related to network failures of our own telemetry posts. Removing the dependency on the bundle reduces the complexity of some distribution use cases like when we embed this as a library in our Unity SDK.

* Silence safe API warnings

* Make access token setable

Previously an access token could only be set once. This changes things so that a new access token value can be set on the events manager and that value is passed through to the api manager where it is actually used.

This makes it so that clients of the library can more easily and explicitly set alternative access tokens without having to hand off via user defaults.

* Make base URL settable

This adds a new public API that makes it easy to set the base URL for where events will be posted.

It was previously possible to override the default URL with a value for MMETelemetryTestServerURL in user defaults. This replaces that backdoor option with a public property on events manager that passes through to the api client where the value is actually used.

Validation for a secure protocol remains but word “test” is not used since this API could be used to point to any alternative backend stack, not just the test server.

Host apps and SDKs can still use user defaults to discover, set, and store the URL for events posting and pass the value through with this new API.

* Remove user defaults hooks for staging testing

This cleans up some references to configuration keys (MMETelemetryTestServerURL and MMETelemetryStagingAccessToken) that used as a backdoor to change the server and access token for testing. There are now actual APIs that replace the old user defaults approach.

This also removes the accessToken getter in API client since that object now always expects to be configured with an access token instead of sometimes pawing through user defaults to find one.

* Always use strong self

* Include date in turnstile guard debug log message

* Remove cart, scan files from dynamic build output

* Omit namespace header for CocoaPods distribution

The namespaced dependencies header file only makes sense to use with custom, vendored integrations in a host SDK or app.