Correct SessionWrapper unit tests
Some of the tests didn't use proper tooling and were all in all a bit messy. 

This commit should clean up the unit tests around the session wrapper and stubs the `handlechallenge` method to ensure that it returns the correct value for unit testing. The test we had before (testing `NSURLSessionAuthChallengeCancelAuthenticationChallenge` may be added back as an integration test). 

LLVM_LTO is also disabled for debug modes. This is to allow breakpoints to operate properly. `-fcxx-modules` and `-fmodules` are added to import #TrustKit into the .mm file.

Co-authored-by: Andrew-Kitchen <andrew.kitchen@mapbox.com>