telemetryMetrics improvements (#113)
* Reformat failedRequests metric 

This should allow the failedRequests metric to be more queryable.

* Track received data 

Implement the event schema changes to begin tracking received data on all requests. 

Tested using breakpoints to track each `data.length` from start to finish.

* Fix bug in event counting

Although this wouldn't be much an issue in our current implementation; once we add in deferred telemetry we would have missed counting any additional batches in APIClient.

Also adds a failed event count if the request is ever nil.

* Implement suggested changes 

Also adds missing tests and removed the synchronous NSURLConnection call that was being used for testing. 

Changed transferred bytes to sent bytes -- this will be reflected in the schema soon.

* Move telemetryMetrics send request method

Moves `telemetryMetrics` send request method to `flush` to have the event send more regularly. 

Previously this method was moved out of `flush` to be more readable but we missed a vital part that `flush` is also called on a timer and `sendTelemetryMetricsEvent` was out of that loop.

* Add telemetryMetrics timer test

Adds a unit test to ensure that we attempt to send a telemetryMetrics event when the event queue timer fires.