Test Failure: Time Zones
Problem:
One of the errors I got once the CI was up and running again was due to time zone difference:

```
MMEDateTests
  testHTTPDateFormatter, (([dateString isEqualToString:@"Sat, 3 Jan 1970 03:28:43 -0800"]) is true) failed - Sat, 3 Jan 1970 11:28:43 +0000
  /Users/vagrant/git/Tests/MMEDateTests.m:18
    NSString *dateString = [MMEDate.HTTPDateFormatter stringFromDate:date];
    XCTAssert([dateString isEqualToString:@"Sat, 3 Jan 1970 03:28:43 -0800"], @"%@", dateString);
}
```

In this case, the test is expecting the string to include a time zone Z of PST. I'm guessing the CI box is somewhere in the UK UTC timezone as bitrise scales which may have revealed this issue.

Options:
1. Force Test to PST to conform to expected string
2. Hard code formatter to UTC allowing for test to play friendly regardless of time zone.
3. Use UTC and drop `Z` completely as it may not be necessary.

Approach:
(2) Opted to maintain the string style output, but set to UTC to use a consistent output of format and timezone representation regardless of which time zone we’re in.

Change:
+ Set formatter to UTC so  tests will be consistently run
+ Include assert comment to log failure date string value