[MobileMetricsManager] Initial Abstraction
Goal:
Provide for the ability to inject a MobileMetricsManager allowing for the ability to test individual components on their responsibilities.

Approach:
Given the creation of a metrics manager in MMEEventManager singleton, inject values on init allowing for the eventual test of individual components. This involved creating instances of MMELogger which can be instantiated from MMEEventsManager and passed in with whichever log level configuration may be preferred.

Changes:
+ MMEMetricsManager with dependencies in init
+ MMELocationManager init with dependencies
+ APIClient init with dependencies
+ Add NSBundle Extension to get MME Bundle
+ Configure each at the root MMEEventsManager