[Client Decoupling] From EventsManager + Metrics
Problem:
Tight coupling can make it painful if not impossible to test and verify behaviors. In order to allow for clearer undestanding of roles/responsibilities, decouple Client from reaching out to global variable EventManager and Metrics/Logger.

Client Should be responsible for wrapping a URL session with the necessary markup to make service api calls. It should not be resopnsible for dictating tracking behaviors/implementations and logging.

Approach:
1. Remove MetricsManager and EventsManager imports to allow for visiblity into dependencies.
2. Add Blocks which can be configured as needed allowing for decoupling of behaviors.
3. Transition call sites to use block syntax rather than reaching out to singletons.
4. Update Testing to use Block params
5. For Client Testing, inject Block Counter to evaluate that the blocks are called where/as many times as expected.