[EventConfigService] Add Config Polling Controller
Problem:
Currently, the client is responsible for API calls, as well as polling and updating our config.  Although this works, it makes testing difficult as well as may lead to developer confusion when looking for what does what.

Goal:
Extract out this responsibility to a service which can be configured and mocked to verify we’re able to handle any setting sent our way.

Approach:
Add MMEConfigService responsible for the polling

Changes:
+ Add MMEConfigService
+ Integrate service with Client API Call
+ Add test validating polling behavior