[Memory Management] Fix Weakself Properties
Problem:
When it comes to escaping blocks, we strive to prevent memory cycles which cause leaks. In doing so weakSelf is included in several places in MME SDK. Unfortunately, although we can message nil with function calls, accessing a property of a nil object will cause a crash.

Why:
Memory Access Error

Approach:
- Use stronSelf + checks on blocks
- Use function accessor which plays friendly with accessor due the ability to message nil.

Changes:
- Stub Protocol Completion Block Handler
- Client Completion Block Handler
- Client Document Note on completion block behavior
-