[MMEPreferences] Abstract Access to NSUserDefaults
Problem:
NSUserDefaults+Category couples both interface and storage. This appraoch has driven the adoption of swapping a static reference to UserDefaults.

Goal:
To allow for clean separation between tests/framework, we need the ability to cleanly separate the interface from the storage so a test does not interfere, or be affected by a host app spinning up it’s own preferences.

Approach: (Class only, not integrated)
Provide Preferences class  as an abstraction around NSUserDefaults that is responsible for providing defaults if values are not set.

Allow for the injection of UserDefaults allowing instances to be created

Changes:
+ Include Public Keys for shared use between NSUserDefaults/Preferences during this transition
+ Create MMEPreferences NSUserDefault Wrapper
+ Build Defaults as part of the accessors so there is no need to register defaults as a two, whitebox knowledge setup