[CertificateRevocationList Test] Model/Filtering
Goal:
Enable testing that was disabled (in MMECertPinTests.m)

Approach:
On review, this test wasn’t about logic for CertPinning. Instead it was around logic that was previously hosted in NSUserDefaults+MMEConfiguration, now MMEPreferences.

To enable test, verify the modeling source was accurate as well as the functional calculation once the filter is applied.

Changes:
+ Add MMEPreferencesTest to include Cert/Hash Logical Checks
- Remove Hash logic checks from CertPinTests as it is a model check rather than a pinning logic check
- rename public/china keys to use prefixes to not collide with UserDefaults