[Client] Refactor Blocks (OnURLResopnse)
Problem:
Given the stagecoach work, it may be ideal to get access to responses for http response feedback.

Additionally, this revealed some redudancies/code duplication which could be shared in a common path.

Approach:
Reaplce onError/onBytes received with onURLResponse which allows for byte size inspection as well as HTTP errors in a single block. While splitting this out it also seemed useful to separate the reporting of HTTP errors from Serialization/Deserialization errors.

Changes:
+ onLogEvent -> onURLResponse() or onSerializationError() with reporting done outside of client
+ Update Client initializers to fall through sharable initializers leading to less code duplication of defaults.
+ Migrate postMetadata to use designated performRequest which automatically tracks HTTPErrors/Response data sizing via blocks.
+ Update BlockCount + ClientTests to reflect
+ Enable the registration/listening of Serialization/HTTErrors from the MMEEventsManager