[MMEEventConfigProviding] Move Public -> Internal
Problem:
MMEEVentConfigProviding was originally puposed to enable read only access of configurations through the project to enable dependency injection of instance dependencies instead of singletons allowing for determinant testing.

Although it’s served it’s purpose it doesn’t need to be exposed publically since consuming SDKs can interact with preferences as they may rightly want to do with a GUI editor.

Expectations of Success:
1. Remove MMEEventConfigProviding from public header space
2. SDK Builds

Approach:
To remove MMEEventConfig from the public header space it needs to be removed from MMEEvent.h imports. To accomplish this, the only model which requires this import (Turnstile) was moved to an internal header (along with designated initializers that don’t need to be exposed to the public)

Changes:
- Remove MMEEventConfig from Public header space
+ Migrate MMEEvent initializers to a private header to allow access internally, but limit visibility externally
+ Update comments to follow objective-c style docs