[MMEUIApplicationWrapper] Tests
Goal:
Boost class coverage to 100%

Approach:
Rather than directly accessing UIApplication.shared, provide mechanism to inject dependency allowing to test this classes responsibilities.

To achieve this, pass in a MMEUIApplicationWrapper, the interface that is captured by this class and the UIApplication. By passing this in, we can ensure our wrapper is making the correct calls to the dependency UIApplication.

Change:
+ Add Docs
+ Add dependency property
+ Add Designated/Default initializers
+ Add Tests to verify calls/responses