iOS (App Extension) Freindliness Continued
Context:
https://github.com/mapbox/mapbox-events-ios/pull/205

Although the above PR allowed for the content scale values to be extracted while in an extension, it did not cover all cases.

MMEUIApplicationWrapper also leverages UIApplication.shared to example application state as well as handle background task communication.

Although this is unavailable in NSExtensions, this can be stubbed out for the time being until a longer term decision is made upon how to handle lengthy actions/collections in an extenion.

For further support:
https://developer.apple.com/library/archive/documentation/General/Conceptual/ExtensibilityPG/ExtensionScenarios.html#//apple_ref/doc/uid/TP40014214-CH21-SW1