[src] Use lazy initializers for MMELocationManager and CLLocationManager
Starting from iOS14 in order to get locationAuthorization(and
accuracyAutorization) we need to call instance methods of
CLLocationManager. Since it is possible for SDKs to send userTurnstile
event before finishing construction of MMEEventManager (we have
`initializtion` callback in `initializeWithAccessToken`) we might end up
in situation when CCLocationManager is not created yet.

To avoid behaviour above we introduce two lazy initializers for:
- MMELocationManager in MMEEventManager
- CCLocationManager in MMELocationManager
so both location managers will be always available whenver we try to
access to them.

After this change there is no need to set CCLocationManager pointer to
nil when we call `stopUpdatingLocation` since location manager now is
always available.